# src/Parbat.Core/Parbat.Core.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Parbat.Core.API",
    "description": "An open source ERP for Universities",
    "contact": {
      "name": "Dr. Muhammad Qasim Pasta",
      "url": "https://twitter.com/mqpasta",
      "email": "qasim.pasta@ieee.org"
    },
    "license": {
      "name": "The MIT License (MIT)",
      "url": "http://opensource.org/licenses/mit-license.php"
    },
    "version": "v2.0"
  },
  "paths": {
...
    "/api/v2.0/Student/Get/{id}": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Get a record",
        "operationId": "StudentGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.0/Student/Update": {
      "put": {
        "tags": [
          "Student"
        ],
        "summary": "Update a student",
        "operationId": "StudentUpdate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/Student/Create": {
      "post": {
        "tags": [
          "Student"
        ],
        "summary": "Create",
        "operationId": "StudentCreate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/Student/Delete/{id}": {
      "delete": {
        "tags": [
          "Student"
        ],
        "summary": "Delete",
        "operationId": "StudentDeleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v2.0/StudentGroup/Get/{id}": {
      "get": {
        "tags": [
          "StudentGroup"
        ],
        "summary": "Get a record",
        "operationId": "StudentGroupGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGroup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGroup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGroup"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.0/StudentGroup/Update": {
      "put": {
        "tags": [
          "StudentGroup"
        ],
        "summary": "Update a studentGroup",
        "operationId": "StudentGroupUpdate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentGroup"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentGroup"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentGroup"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGroup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGroup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGroup"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/StudentGroup/Create": {
      "post": {
        "tags": [
          "StudentGroup"
        ],
        "summary": "Create",
        "operationId": "StudentGroupCreate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentGroup"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentGroup"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentGroup"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGroup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGroup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGroup"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/StudentGroup/Delete/{id}": {
      "delete": {
        "tags": [
          "StudentGroup"
        ],
        "summary": "Delete",
        "operationId": "StudentGroupDeleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v2.0/StudentProgram/Get/{id}": {
      "get": {
        "tags": [
          "StudentProgram"
        ],
        "summary": "Get a record",
        "operationId": "StudentProgramGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgram"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgram"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgram"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.0/StudentProgram/Update": {
      "put": {
        "tags": [
          "StudentProgram"
        ],
        "summary": "Update a StudentProgram",
        "operationId": "StudentProgramUpdate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentProgram"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentProgram"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentProgram"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgram"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgram"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgram"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/StudentProgram/Create": {
      "post": {
        "tags": [
          "StudentProgram"
        ],
        "summary": "Create",
        "operationId": "StudentProgramCreate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentProgram"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentProgram"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentProgram"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgram"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgram"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgram"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/StudentProgram/Delete/{id}": {
      "delete": {
        "tags": [
          "StudentProgram"
        ],
        "summary": "Delete",
        "operationId": "StudentProgramDeleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Student": {
        "required": [
          "LastName",
          "Name"
        ],
        "type": "object",
        "properties": {
          "StudentID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "Identifier": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "minLength": 1,
            "type": "string"
          },
          "LastName": {
            "minLength": 1,
            "type": "string"
          },
          "StudetGroupID": {
            "type": "string",
            "nullable": true
          },
          "Key": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "StudentGroup": {
        "required": [
          "Name",
          "ShortName"
        ],
        "type": "object",
        "properties": {
          "StudentGroupID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "ShortName": {
            "minLength": 1,
            "type": "string"
          },
          "Name": {
            "minLength": 1,
            "type": "string"
          },
          "IsActive": {
            "type": "boolean"
          },
          "Key": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "StudentProgram": {
        "required": [
          "StudentID",
          "StudentStatusID"
        ],
        "type": "object",
        "properties": {
          "StudentProgramID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "StudentID": {
            "type": "integer",
            "format": "int64"
          },
          "BatchID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "StartDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "EndDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "StudentStatusID": {
            "type": "integer",
            "format": "int64"
          },
          "OverrideCurriculumID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "MinorCurriculumID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "StudentGroupID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "Key": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}